The Android SDK should not be required when building a pure Java project.
Summary:
Instead of throwing `NoAndroidSdkException` as part of the
buildup to create an `ExecutionContext`, only throw
`NoAndroidSdkException` when `ExecutionContext.getAndroidPlatformTarget()` is
invoked and the Android SDK has not been specified by the user.
This way, a `Build` can create an `ExecutionContext` and run its course
without the creation of an `AndroidPlatformTarget` as a requirement.
This will make it easier to use Buck for a non-Android Java project.

Also changed `NoAndroidSdkException` to be an unchecked exception,
as it is not recoverable. This helps eliminate a number of catch blocks
in the code.

See https://github.com/facebook/buck/issues/7.

Test Plan:
As follows:
    unset ANDROID_HOME
    unset ANDROID_SDK
    rm -rf local.properties
    buck build buck