Push buck into headless mode.
Summary:
With the change to use jython, _something_ is causing AWT to be initialized. This is causing buck to appear as an AWT app, with its own menu bar in OS X. Stop that by indicating that we're headless.

Note: this will cause problems if we ever do anything using AWT, so we might want this to be an option in a later diff. Having said that: who uses the AWT these days?

Test Plan:
buck build buck
Watch as nothing untoward happens