Running an sh_test() was throwing an exception if it had been built but not executed.
Summary:
Refactored some of the test rule code so that it must check whether its test
results have been cached when considering whether the test needs to be run.

Ideally, TestRule would be an abstract class rather than an interface so that
we could standardize some behavior across classes that implement TestRule.

Test Plan:
Ran the following:

  TESTS=`buck targets --type sh_test | xargs`
  buck clean
  buck build $TESTS
  buck test $TESTS

Previously, this failed in fbandroid because `buck test` would see that the rule
was cached, so it assumed that the test results file was also written.
This was an error that was fixed in java_test(), but not sh_test().