Fix 'buck project' for annotation processing
Summary:
D766822 includes an AndroidLibraryRule.Builder that has its own
AnnotationProcessingParams.Builder rather than using the inherited one
from JavaLibraryRule.Builder.  This breaks 'buck project's ability to
generate IntelliJ files that work with annotation processing.

This change fixes that and adds a unit test to verify that this does
not break in the future.

Test Plan:
- unit tests
- 'buck project' in fb4a and verify module_...iml references annotation
  sources.