Remove the uncachedBuildRules field from AndroidTransitiveDependencies.
Summary:
The way we were creating this field does not make sense with respect to
our upcoming changes about what it means for a rule to be "cached."
It also required Optional.<BuildContext>absent() to be passed in many
cases, which does not make for a reliable API, as that means that the
uncachedBuildRules field in AndroidTransitiveDependencies was unreliable.

Test Plan: Sandcastle builds.