Pass the project root as a parameter to Main.runMainWithExitCode().
Summary:
This fixes things so that the root of the project is taken as a parameter
when running Buck programmatically. This will make it possible to run
Buck in different project directories for integration tests.

Recently, all of the classes that extend AbstractCommandRunner were updated
to take an ArtifactCache as a parameter. This diff replaces that with a
more extensible CommandRunnerParams type because updating all of the call
sites can be tedious, so now we can introduce new parameters to
AbstractCommandRunner in one place. Currently, CommandRunnerParams has two
fields: ArtifactCache and ProjectFilesystem.

Test Plan:
Sandcastle builds.
Verify that `git grep 'new File(".")`` returns only one result under src/.