Buck daemon all the commands
Summary:
This looks like a big change, but it's just several small changes and some wiring. Main now creates a
single static Parser which persists for the potentially long lived Buck process and is passed to every command.
Additionally, if Buck is running as a daemon, Main watches the file system and subscribes the Parser to file
change events. The Parser interface is tightened up so that it either reads all of the build files in the
project and caches them (so doesn't need to read any future build files) or the build files for a given target
(in which case a subsequent call to read all of the files will replace the subset of cached files for that
target). If the Parser is notified of a change to a non-source file then the cache will be invalidated and
the build files parsed again on subsequent calls. The static BuildFileToJsonParser methods have been moved to
a new ProjectBuildFileParser class so that a instance of this object can be passed to the Parser to use when it
needs to parse build files. This allows the cache invalidation logic in Parser to be tested using a fake
ProjectBuildFileParser which can count the number of times the Parser has parsed build files in a number of
different scenarios.

Test Plan:
1) buckd
2) time buck test --all
3) time buck test --all
4) check that the time reported in step 3) is significantly lower than in step 2)
5) repeat steps 2-4 with other buck commands