Revert logic that printed out when tests stop and start.
Summary:
This was a hack introduced by a previous commit.
It was used to help determine which test was running forever on our
continuous integration builds. Now that we have the ability to set
a default timeout, this is no longer necessary. Deleting this code
will remove a considerable amount of clutter from `--verbose 2`.

Test Plan: Sandcastle builds.