Replace Map<String, BuildRule> with an extensible BuildRuleBuilderParams.
Summary:
Historically, we kept track of build rules as we built them in a
Map<String, BuildRule>. This also doubled as the parameter that was passed to
BuildRuleBuilder.build().

This diff introduces BuildRuleBuilderParams instead of the Map. This is
an important improvement because:

* The raw Map is no longer exposed: it is append-only from the client's perspective.
* We can now add more parameters to BuildRuleBuilder.build() without updating
  all of the call sites.
* In tests, we would frequently create a BuildRule programmatically in one
  statement and then add it to the Map in a second statement. Now this can
  be done in a single statement using BuildRuleBuilderParams.buildAndAddToIndex().

Much of the s/Map<String, BuildRule>/BuildRuleBuilderParams/g was done using
Eclipse for refactoring, so unfortunately, this may have introduced some
long-line violations or other formatting anomalies.

Another thing that magically fell out of this diff was the generification of
BuildRuleFactory and BuildRuleBuilder. Without this,
BuildRuleBuilderParams.buildAndAddToIndex() would return an ordinary BuildRule,
so using it would have required a ton of casting. I didn't seem to get all of
the generics quite right in the case of subclasses, so casting is necessary
for AndroidLibaryRule, RobolectricTestRule, and ApkGenrule. I'm hoping that
someone who is more knowledgeable about generics can help with these three cases
in a follow-up diff.

Test Plan: Sandcastle builds.