[buck] Add an option for native libraries to be treated as assets
Summary:
This makes it so that any native library declared using `ndk_build` or
`prebuilt_native_library` can indicate if it needs to be treated as an asset.
This could be useful if you want to extract/load these libraries on demand
and using mechanisms other than the Android system library loader.

Moreover, it will respect the build `cpu_filter`
and pick up the appropriate library for that cpu arch.

Libraries using this option are placed under `/assets/lib/` inside the APK.