Add BuildRuleParams#getPathRelativizer().
Summary:
This makes it possible to relativize paths in `AbstractBuildRuleBuilder#getInputs()`.
This will help with implementing new integration tests, as we have places where
we have failed to use a ProjectFilesystem to resolve relative paths. Because our
integration tests run outside of the root directory, this causes problems that today
are observed only in integration tests.

Test Plan: Sandcastle builds.