Pass ProjectFilesystem into BuckConfig#createArtifactCache().
Summary:
This is a temporary workaround to pass a ProjectFilesystem into
BuckConfig#createArtifactCache() so that relative paths to a directory
cache get resolved correctly.

Ideally, ProjectFilesystem would be injected into BuckConfig's constructor
so that this would not be necessary. Making that change would be an easy fix
for BuckConfig, but it turns out to be quite a bit of work in terms of
updating BuckConfigTest.

I will file a task to do this the right way, but for now, this is on my
critical path for the ABI work, so I need to get unblocked.

Test Plan: Sandcastle builds.