Add unit test for BuildContext#logBuildInfo().
Summary:
Ironically, with this diff, EMMA claims that our code coverage for
BuildContext.java by lines is "100% (70.9/71)", but our code coverage
by classes (which usually means internal or anonymous classes) is
only "80% (4/5)". I'm struggling to figure out where the untested
inner class is. According to EMMA, the only partially tested line is:

    Preconditions.checkState(!bootclasspathEntries.isEmpty(), <msg>)

presumably because whenever it is executed,
`!bootclasspathEntries.isEmpty()` is true. This happens to occur in
an anonymous `Supplier`.

Test Plan: Sandcastle builds.