Remove dependency on guava from the AbiWriter.
Summary:
It would be useful to use the AbiWriter as an annotation processor without requiring it
be compiled directly into buck (which also allows us to use other processors at the same
time) The code isn't terribly complex, so removing dependencies on anything other than
the JDK will work. That means that the single jar generated by buck needs to be included
when using the AbiWriter.

Test Plan: buck test --all