Modify a number of ShellSteps to use relative paths exclusively.
Summary:
Because these shell steps now take relative paths,
they can be resolved to absolute paths appropriately in
`getShellCommandInternal(ExecutionContext)` by
`ExecutionContext.getProjectFilesystem().getPathRelativizer().apply()`.
If we end up running `ShellStep`s on remote machines in the
future, this will make it easier to serialize a `ShellStep` and run it
on the foreign machine.
This also helps with integration tests that are run in [foreign] temp directories.

Test Plan: Sandcastle builds.