Fix a regression in `buck audit owner` by relativizing file paths properly.
Summary:
This regression was introduced in b592dc1c1c01ffba5b72a3ecae202f7f2f10511c.

That diff introduced a change where paths were relativized before being
passed to the constructor of an `InputRule` in `AbstractCachingBuildRule.getInputs()`.
Unfortunately, `AuditOwnerCommand` was still using the `InputRule(String)` constructor
for an `InputRule`, which created an `InputRule` using `new File(String)`, which
means the `File` created for an `InputRule` in `AuditOwnerCommand` did not go through
the same path relativization as those created by an
`AbstractCachingBuildRule.getInputs()`. This mismatch caused the regression in
`buck audit owner`.

This diff eliminates the `InputRule(String)` constructor and forces `InputRule`s to
be created via a static method that requires a `pathRelativizer` as a parameter.
This also introduces `InputRuleTest` that ensures that the `BuildTarget` for an
`InputRule` is composed of only the relative path.

Test Plan: Sandcastle builds.