Update DefaultJavaLibraryRule so it implements AbiRule.
Summary:
When a DefaultJavaLibraryRule is built, it writes its RuleKey and ABI-key-without-deps
to disk. This makes it straightforward to extend DefaultJavaLibraryRule so that it
implements AbiRule.

Methods of AbiRule that require reading a file from disk are now passed a ProjectFilesystem.

This also fixes an issue in AbstractCachingBuildRule where a RuleKey was not written
when a recompile was avoided due to unchanged ABI keys.

This fixes a bug in ProjectFilesystem, as well as introduces some new methods.
Unit tests are added, as appropriate.

Test Plan:
Added a new test to DefaultJavaLibraryRuleIntegrationTest to verify
that ABIs get written and that they are used successfully to avoid
unnecessary recompilation.
Sandcastle builds.