Ensure System.exit is called even if we throw an exception all the way to main.
Summary:
Encountered an issue in practice while re-working buck's interaction
with buck.py whereby an IOException (part of the checked signature of
most of the main() machinery) would cause buck to stall waiting on
non-daemon threads to die that are never going to.

There are a variety of reasons why IOException should not have been
allowed to go all the way to main, but regardless those bugs should be
caught more gracefully so that we print the stack trace (just like we
would've before) and exit the process so that the user is not confused
about the state of the process.

Test Plan:
Modify the code to throw an arbitrary IOException and verify that the
process now stops as expected.