Fork to invoke the interpreter and switch to using python by default (parse perf pt3)
Summary:
Once again invoke python/jython as a separate process.  Python is selected
as the new default for performance reasons and because this manner of
invoking Jython created a subtle regression in buck's own unit tests
(because they create many instances of the parser for simple jobs and
Jython's startup cost is so enormous).  Note that with buckd, many parse
invocations may yield slightly worse performance but in practice because
python starts up so much faster it should not be an issue in practice
(python buck.py --server starts up in about 0.1s whereas Jython takes about 3s).

Also had to upgrade jython to 2.5.4-rc1 due to upstream issues #1962
and #1972 to keep it working properly in this environment.

Test Plan:
Used buck build and buck targets to verify that both ./bin/jython and
python interpreters work.

Parse errors produce suboptimal results but are detected.