Lazily construct buildFiles.
Summary:
Maybe this should be abstracted into a Cache class that takes a Supplier
and caches the last returned value until you invalidate it.

Only one call to constructBuildFileTree, and it's fast-ish.

The time also moved from watchFileSystem to parseBuildFilesForTargets,
which means that it will show up in the Chrome trace, which is good.

Test Plan: Needs tests.