Introduce a Buildable that encapsulates the meat of a build rule.
Summary:
Both Michael and I have demonstrated that attempting to introduce a brand new wrapper
around building a rule and caching its output is prone to being hard to land. Take a
more incremental approach by introducing an interface that things that can be built
should implement.

The next step will be to make the Buildable a constructor arg for the
AbstractCachingBuildRule.

Test Plan: buck test --all