Remove OutputKey.
Summary:
OutputKey is not currently used, nor is it tested.
I'm not convinced that it is a concept that we should have,
so let's remove it for now.

I also want to get this out of the way because I want to
do away with the idea of non-idempotent RuleKeys.

Test Plan: Sandcastle builds.