Let Buildable.appendToRuleKey() throw IOException.
Summary:
In an upcoming build, I am going to change the behavior of
RuleKey so that it no longer catches an IOException and
flags a RuleKey as non-idempotent. Instead, it will just fail.

A side-effect of this is that many method signatures will have
to be updated with "throws IOException", so I did not want to
do all of that in one diff, since it would mix a bunch of
mechanical changes (that should be easy to review) with
substantive changes (that should be the focus of the review).

Test Plan: Sandcastle builds.