Fire a log event when we spot duplicate files/directories when building an uberjar.
Summary:
First we failed completely for the duplicate entries,
and we said nothing (other than catastrophic build failure).
Then we succeeded silently, dropping the duplicate entries,
and we said nothing.
Then we fired log events when we saw duplicate entries,
and so we say something.

This is in response to a request on the buck mailing lists to make it visible
when duplicate files/directories are encountered when building an uberjar.

Test Plan: buck test --all