Second run at handling creation of symlinks on Windows.
Summary:
Creating symlinks on Windows is _hard_. Files.createLink only works on files, but it's possible
for us to get given a directory to symlink. Handling this on XP requires escalated privileges
and isn't recommend. With this diff we:

* Create a symlink directly if we're not on Windows.
* If we're on Windows and we're linking a file, create a link
* Otherwise, recursively copy the directory.

That last step means that linking to a directory will be very inefficient on Windows.

Test Plan: buck test --all