Push Buildable interface down to child classes of AbstractCachingBuildRule.
Summary:
Next step in the plugin work. We would like the AbstractCachingBuildRule to ultimately
be a wrapper around Buildable instances. In order to do this, things that currently
extend it need to implement the newly introduced Buildable interface.

We could attempt to mash the change where the Buildable is a field in the
AbstractCachingBuildRule but that turns out to be a wide ranging change that impacts
a lot of code. In addition, we still need a way to annotate Buildables as being
either TestRules or AndroidLibraries.

Test Plan: buck test --all