Introduce DebuggableTemporaryFolder.
Summary:
Frequently, when debugging an integration test,
I do not want the TemporaryFolder to delete itself at
the end of the test because I want to be able to inspect
its contents. A DebuggableTemporaryFolder provides the
option to do just that.

This also updates ProjectWorkspace to require a
DebuggableTemporaryFolder so that developers will start
using this by default.

Test Plan: Sandcastle builds.