Fix some build cache interaction issues in AndroidResourceRule.
Summary:
(1) `AndroidResourceRule` was writing its output to `buck-out/bin`
instead of `buck-out/gen`. This was causing the following check
in `AbstractCachingBuildRule` to fail:

```
if (pathToOutputFile != null && pathToOutputFile.startsWith(BuckConstant.GEN_DIR)) {
  String prefix = BuckConstant.GEN_DIR + '/' + getBuildTarget().getBasePathWithSlash();
  buildInfoRecorder.recordArtifact(pathToOutputFile.substring(prefix.length()));
}
```

After fixing this for `AndroidResourceRule`, I manually inspected the
`getPathToOutputFile()` method in all build rules, and verified that
if a non-null value was returned, it started with `buck-out/gen`.

(2) `AndroidResourceRule` had two fields that were not being included in its RuleKey.

Test Plan: Sandcastle builds.