Be smarter about preventing Cassandra failures from being too chatty.
Summary:
Funnel all ConnectionExceptions through one place and print the first 10
of them. 1 seems like too few if the connection is flaky; infinity seems
like too many if the user is developing offline.

Test Plan:
This was hard to test as jmorrow was fixing Cassandra as I was writing
this, so I was seeing fewer failures over time.