Unify the treatment of stdout and stderr in ProcessExecutor.
Summary:
This changes the behavior so that, for a ShellStep, each of stdout and stderr
will either be printed or captured (these states may be set independently).
If not printed, we always capture so that if the ShellStep fails, we can
always print what was captured in the event that it was not streamed directly
to stdXXX.

Test Plan: Sandcastle builds.