Introduce our own implementation of a ZipOutputStream.
Summary:
We currenty depend on the command line zip command in order to create zip files. This means that we cannot
currently build an APK on Windows, which is obviously suboptimal. In the ideal world, we'd use something
that ships with the JRE, but it turns out that the two obvious candidates fail. Our requirements are:

1) Can support changing the deflation level per-file.
2) Can support writing the same file more than once to the zip, overwriting existing entries.
3) As 2, but appending the new entry.

ZipOutputStream meets 1, but 2 or 3. ZipFileSystem supports 2, but not 1 or 3. The classes in this
diff allow us to support all three options.

Test Plan: buck test --all