Propagate rule keys from annotation processors to dependent rules.
Summary:
Currently we do not propagate the rule key from annotation processors
to dependent rules.  This means if you change an annotation processor,
buck will consider cached entries for rules that depend on the processor
to still be valid.

This change adds support to AnnotationProcessingData/Params to transmit
a list of RuleKey from the annotation processing rule to the dependent
rule.