Fix python_library so it caches correctly.
Summary:
Previously, if you had a python_binary rule that depended on a
python_library rule, and both rules were cached in cassandra,
there would be an issue executing the python_binary rule
(presumably as part of a genrule) because the files from the
python_library rule would not be in buck-out/bin/ where the
python_binary rule expected them to be.

This changes things so that the python_library writes its files
to buck-out/gen/, and those files are archived with the python_library
rule such that when it is pulled down from cassandra, it is
unpacked in the way that python_binary needs.

Test Plan: Sandcastle builds.