Use a LogEvent rather than writing to stderr directly.
Summary:
Whenever we write to stderr, we stop writing to the
SuperConsole, so unless an error is particularly egregious,
we should use the BuckEventBus's facility for logging so
an error is displayed below the SuperConsole without killing it.

See DirtyPrintStreamDecorator for details about why writing
to stderr kills the SuperConsole.

Test Plan: Sandcastle builds.