Modify PrebuiltJarRule so that its ABI is the hash of the binary_jar file.
Summary:
Note that when the `binary_jar` of a `prebuilt_jar` is a `genfile`,
the ABI is not guaranteed to be consistent across machines.
The end-user must be responsible for generating the JAR file in a
deterministic way to guarantee that the ABI will be reliable.

This could potentially be avoided (or at least diminished) if the
ABI were a hash of the contents of the .class files contained in the
JAR instead.

Test Plan: Sandcastle builds.