Merge string resources for region specific locales.
Summary:
Merge region specific locale resources with the corresponding base locale resources.
For example, if there are separate string resources in an android project for locale
"es" and "es_US", when an application running on a device with locale set to "Spanish
(United States)" requests for a string, the Android runtime first looks for the string
in "es_US" set of resources, and if not found, returns the resource from the "es" set.
We merge these because we want the individual string json files to be self contained
for simplicity.