Cache InputFile content hashes.
Summary:
When run as buckd a cache of the hashes of InputFile contents
is maintained, avoiding the need to read and hash a potentially large
number of files each build to determine whether they have
changed. Hashes are invalidated when the corresponding file is changed
and buckd is notified via the file system watcher.

Test Plan:
1) buck test --all
2) buckd
3) buck build buck
4) buck build buck
5) check that the durations of the parse and build phases are both < 0.1 second