Make MoreFutures.isSuccess not throw when it is interrupted.
Summary:
It's possible for a future (like Guava's AbstractFuture) to return true
from isDone but still throw an InterruptedException from get() if the
current thread has been interrupted.  The old code threw a
RuntimeException in this case, which was unnecessary.  Change it to just
retry and reset the thread interrupt status instead.

This solution was informed by the "Noncancelable tasks" section of
http://www.ibm.com/developerworks/java/library/j-jtp05236/index.html

Test Plan:
Unit tests.  The new test failed 5 out of 5 times on the old code on my
machine, but it could pass spuriously (without the doNotOptimizeMe loop,
it was passing about 50% of the time).  With the new code, it should
pass deterministically.