Custom file format for string files.
Summary:
We were using JSON as the file format for strings dumped as assets. The
resulting `APK` for fb4a is around 400-500 KB bigger in size, which is not
acceptable. This introduces a custom format (explained in javadocs).

In order to further trim the file size, we store resource ids instead of
resource names. In order to achieve this, the strings file generation step is
now delayed, running it after the final `R.java` file is created by `aapt`.
Instead of parsing `R.java`, I also make `aapt` generate `R.txt` which is very
easy to parse.